tests/requirements.txt
pytest>=7.0
Pillow>=9.0
